# project_back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Exercices",
        "version": "1.0"
    },
    "paths": {
        "/exercices": {
            "get": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Liste tous les exercices",
                "operationId": "e558c39539d68edb9fdec37196f9aebf",
                "responses": {
                    "200": {
                        "description": "Liste des exercices récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Exercice"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Créer un nouvel exercice",
                "operationId": "c221ea923daadc2bbcdda1dd62d9a457",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pompes"
                                    },
                                    "advice": {
                                        "type": "string",
                                        "example": "Gardez le dos droit"
                                    },
                                    "series": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "repetitions": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "id_training": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exercice créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Séance non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Séance non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/exercices/{id}": {
            "get": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Affiche les détails d'un exercice",
                "operationId": "5941d5da0480f89b22368448de8b23b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'exercice",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'exercice récupérés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exercice non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercice non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Mettre à jour un exercice",
                "operationId": "c9b41cb1bb4fcbb1ab615203aa90e53f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'exercice",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pompes"
                                    },
                                    "advice": {
                                        "type": "string",
                                        "example": "Gardez le dos droit"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercice mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exercice non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercice non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Supprimer un exercice",
                "operationId": "dda7c2d1fdbfa0a4b9b93e4a64c86fdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'exercice",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercice supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercice supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exercice non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercice non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Enregistrer un utilisateur",
                "description": "Crée un nouvel utilisateur.",
                "operationId": "f1fde7f4a932e393f6042d1b92e65b0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Connexion d'un utilisateur",
                "description": "Authentifie un utilisateur et génère un jeton d'authentification.",
                "operationId": "71d8adfdcf176b9e5ef0ade00011ce34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "user": {
                                        "id": 1,
                                        "name": "John Doe",
                                        "email": "john.doe@example.com"
                                    },
                                    "token": "generated_token"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Exercice": {
                "title": "Exercice",
                "description": "Un modèle représentant un exercice d'entraînement",
                "properties": {
                    "id": {
                        "description": "ID de l'exercice",
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "description": "Nom de l'exercice",
                        "type": "string",
                        "example": "Pompes"
                    },
                    "description": {
                        "description": "Description de l'exercice",
                        "type": "string",
                        "example": "Exercice de renforcement musculaire"
                    },
                    "series": {
                        "description": "Nombre de séries",
                        "type": "integer",
                        "example": 3
                    },
                    "repetitions": {
                        "description": "Nombre de répétitions",
                        "type": "integer",
                        "example": 12
                    },
                    "training_id": {
                        "description": "ID de la séance associée à cet exercice",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Date de création de l'exercice",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour de l'exercice",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Training": {
                "title": "Training",
                "description": "Un modèle représentant une séance d'entraînement",
                "properties": {
                    "id": {
                        "description": "ID de la séance",
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "description": "Nom de la séance",
                        "type": "string",
                        "example": "Séance de musculation"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur propriétaire de la séance",
                        "type": "integer",
                        "example": 1
                    },
                    "exercices": {
                        "description": "Liste des exercices associés à la séance",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Exercice"
                        }
                    },
                    "created_at": {
                        "description": "Date de création de la séance",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour de la séance",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Modèle représentant un utilisateur",
                "properties": {
                    "id": {
                        "description": "ID de l'utilisateur",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de l'utilisateur",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Adresse email de l'utilisateur",
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "description": "Rôle de l'utilisateur",
                        "type": "string",
                        "example": "admin"
                    },
                    "picture": {
                        "description": "URL de la photo de profil de l'utilisateur",
                        "type": "string",
                        "example": "https://example.com/profile.jpg"
                    },
                    "email_verified_at": {
                        "description": "Date de vérification de l'email",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Date de création de l'utilisateur",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour de l'utilisateur",
                        "type": "string",
                        "format": "date-time"
                    },
                    "trainings": {
                        "description": "Liste des séances d'entraînement associées à cet utilisateur",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Training"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}